#### Problem
It can possible for the capacity of the data structure to change if elements are added while it is full in the loop.
#### Summary of Changes
Compare to constant vs `VecDeque::capacity()`

## Benchmarks
### Results (release mode):
```
old 5000: 114.716µs
new 5000: 69.948µs
```
```
old 100000: 1.222032ms
new 100000: 939.642µs
```

### Code:
```rust
    #[test]
    fn insert_bundle() {
        let n = 5_000;
        let mut deque = VecDeque::with_capacity(n);
        let deserialized_bundles = (0..n)
            .map(|i| {
                let tx = test_tx();
                let tx_bytes = serialize(&tx).unwrap();
                let mut packet = Packet::from_data(None, tx.clone()).unwrap();
                ImmutableDeserializedBundle::new(
                    &mut packet_bundle::PacketBundle {
                        batch: solana_perf::packet::PacketBatch::new(
                            iter::repeat(packet).take(4).collect_vec(),
                        ),
                        bundle_id: String::from("34888034-9518-11ee-b9d1-0242ac120002"),
                    },
                    None,
                )
                .unwrap()
            })
            .collect_vec();

        let start = Instant::now();
        BundleStorage::insert_bundles_old(&mut deque, deserialized_bundles, true);
        println!("old {n}: {:?}", start.elapsed());

        let mut deque = VecDeque::with_capacity(n);
        let deserialized_bundles = (0..n)
            .map(|i| {
                let tx = test_tx();
                let tx_bytes = serialize(&tx).unwrap();
                let mut packet = Packet::from_data(None, tx.clone()).unwrap();
                ImmutableDeserializedBundle::new(
                    &mut packet_bundle::PacketBundle {
                        batch: solana_perf::packet::PacketBatch::new(
                            iter::repeat(packet).take(4).collect_vec(),
                        ),
                        bundle_id: String::from("34888034-9518-11ee-b9d1-0242ac120002"),
                    },
                    None,
                )
                .unwrap()
            })
            .collect_vec();

        let start = Instant::now();
        BundleStorage::insert_bundles(&mut deque, deserialized_bundles, true);
        println!("new {n}: {:?}", start.elapsed());
    }
```